stream/tcp: make RandomGetWrap conversions explicit
Same pattern as the htp copy of RandomGetWrap: RandomGet returns a
signed long that trips -fsanitize=implicit-conversion when assigned to
the unsigned accumulator. Cast explicitly.

+++ src/stream-tcp.c
@@ -467,10 +467,10 @@ static int RandomGetWrap(void)
     unsigned long r;
 
     do {
-        r = RandomGet();
+        r = (unsigned long)RandomGet();
     } while(r >= ULONG_MAX - (ULONG_MAX % RAND_MAX));
 
-    return r % RAND_MAX;
+    return (int)(r % RAND_MAX);
 }
 
 static const char *UrgentPolicyToString(enum TcpStreamUrgentHandling pol)
